chore: Let cargo test work out of the box
