+++ check/Cargo.toml
@@ -32,11 +32,9 @@ gluon_parser = { path = "../parser", version = "0.9.4", optional = true } # GLUO
 [dev-dependencies]
 env_logger = "0.6"
 
+gluon_parser = { path = "../parser", version = "0.9.4" } # GLUON
 gluon_format = { path = "../format", version = ">=0.9" }
 
 collect-mac = "0.1.0"
 pretty_assertions = "0.5"
-
-[features]
-test = ["gluon_parser"]
