+++ parser/tests/indentation.rs
@@ -6,8 +6,8 @@ extern crate gluon_parser as parser;
 mod support;
 
 use base::ast::*;
-use parser::{parse_string, ParseErrors};
-use support::MockEnv;
+use parser::ParseErrors;
+use support::*;
 
 fn parse(text: &str) -> Result<SpannedExpr<String>, ParseErrors> {
     parse_string(&mut MockEnv::new(), text).map_err(|(_, err)| err)
