+++ vm/src/core/optimize.rs
@@ -321,7 +321,7 @@ where
     })
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "test"))]
 mod tests {
     use super::*;
